requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
dj-database-url==0.5.0
django-dotenv==1.4.2
django==2.1.4
pillow==5.3.0
psycopg2-binary==2.7.6.1
pytz==2018.7              # via django
whitenoise==4.1.2
